Changed radar. Obstacles now green. Health working

+++ thegame/thegame/MoveableEntity.cs
@@ -22,7 +22,7 @@ namespace thegame
         protected float spinY;
         protected float spinZ;
         protected bool alive;
-        protected float health;
+        protected int health;
 
         protected static int score;
         public void Score()
